# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Tripline API
  version: 0.1.0
  description: "Self-hosted dependency drift monitor. Authenticate with the session cookie (browser; state-changing requests need header `X-Tripline-CSRF: 1`) or a read-only bearer API token."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tripline_session
    token:
      type: http
      scheme: bearer
      description: Read-only API token (tl_…)
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details: {}
      title: ErrorResponse
paths:
...
  /api/v1/drift:
    get:
      summary: List drift events, newest first
      tags:
        - drift
      parameters:
        - schema:
            enum:
              - open
              - accepted
              - dismissed
          in: query
          name: status
          required: false
        - schema:
            type: string
            format: uuid
          in: query
          name: monitorId
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          in: query
          name: limit
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - events
                properties:
                  events:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - monitorId
                        - monitorName
                        - monitorKind
                        - detectedAt
                        - lastSeenAt
                        - occurrences
                        - severity
                        - changes
                        - status
                        - decidedBy
                        - decidedAt
                      properties:
                        id:
                          type: string
                          format: uuid
                        monitorId:
                          type: string
                          format: uuid
                        monitorName:
                          type: string
                        monitorKind:
                          enum:
                            - http
                            - llm
                            - mcp
                        detectedAt:
                          type: string
                          format: date-time
                        lastSeenAt:
                          type: string
                          format: date-time
                        occurrences:
                          type: integer
                        severity:
                          enum:
                            - breaking
                            - warning
                            - info
                        changes:
                          type: array
                          items:
                            type: object
                            required:
                              - kind
                              - path
                              - severity
                              - message
                            properties:
                              kind:
                                type: string
                              path:
                                type: string
                              severity:
                                enum:
                                  - breaking
                                  - warning
                                  - info
                              before:
                                type: string
                              after:
                                type: string
                              message:
                                type: string
                        status:
                          enum:
                            - open
                            - accepted
                            - dismissed
                        decidedBy:
                          anyOf:
                            - type: string
                              format: uuid
                            - type: "null"
                        decidedAt:
                          anyOf:
                            - type: string
                              format: date-time
                            - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/drift/{id}:
    get:
      summary: Get a drift event
      tags:
        - drift
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - event
                properties:
                  event:
                    type: object
                    required:
                      - id
                      - monitorId
                      - monitorName
                      - monitorKind
                      - detectedAt
                      - lastSeenAt
                      - occurrences
                      - severity
                      - changes
                      - status
                      - decidedBy
                      - decidedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      monitorId:
                        type: string
                        format: uuid
                      monitorName:
                        type: string
                      monitorKind:
                        enum:
                          - http
                          - llm
                          - mcp
                      detectedAt:
                        type: string
                        format: date-time
                      lastSeenAt:
                        type: string
                        format: date-time
                      occurrences:
                        type: integer
                      severity:
                        enum:
                          - breaking
                          - warning
                          - info
                      changes:
                        type: array
                        items:
                          type: object
                          required:
                            - kind
                            - path
                            - severity
                            - message
                          properties:
                            kind:
                              type: string
                            path:
                              type: string
                            severity:
                              enum:
                                - breaking
                                - warning
                                - info
                            before:
                              type: string
                            after:
                              type: string
                            message:
                              type: string
                      status:
                        enum:
                          - open
                          - accepted
                          - dismissed
                      decidedBy:
                        anyOf:
                          - type: string
                            format: uuid
                          - type: "null"
                      decidedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/drift/{id}/accept:
    post:
      summary: "Accept: the observed structure becomes the baseline"
      tags:
        - drift
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - event
                properties:
                  event:
                    type: object
                    required:
                      - id
                      - monitorId
                      - monitorName
                      - monitorKind
                      - detectedAt
                      - lastSeenAt
                      - occurrences
                      - severity
                      - changes
                      - status
                      - decidedBy
                      - decidedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      monitorId:
                        type: string
                        format: uuid
                      monitorName:
                        type: string
                      monitorKind:
                        enum:
                          - http
                          - llm
                          - mcp
                      detectedAt:
                        type: string
                        format: date-time
                      lastSeenAt:
                        type: string
                        format: date-time
                      occurrences:
                        type: integer
                      severity:
                        enum:
                          - breaking
                          - warning
                          - info
                      changes:
                        type: array
                        items:
                          type: object
                          required:
                            - kind
                            - path
                            - severity
                            - message
                          properties:
                            kind:
                              type: string
                            path:
                              type: string
                            severity:
                              enum:
                                - breaking
                                - warning
                                - info
                            before:
                              type: string
                            after:
                              type: string
                            message:
                              type: string
                      status:
                        enum:
                          - open
                          - accepted
                          - dismissed
                      decidedBy:
                        anyOf:
                          - type: string
                            format: uuid
                          - type: "null"
                      decidedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/drift/{id}/dismiss:
    post:
      summary: "Dismiss: mute this exact set of changes"
      tags:
        - drift
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - event
                properties:
                  event:
                    type: object
                    required:
                      - id
                      - monitorId
                      - monitorName
                      - monitorKind
                      - detectedAt
                      - lastSeenAt
                      - occurrences
                      - severity
                      - changes
                      - status
                      - decidedBy
                      - decidedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      monitorId:
                        type: string
                        format: uuid
                      monitorName:
                        type: string
                      monitorKind:
                        enum:
                          - http
                          - llm
                          - mcp
                      detectedAt:
                        type: string
                        format: date-time
                      lastSeenAt:
                        type: string
                        format: date-time
                      occurrences:
                        type: integer
                      severity:
                        enum:
                          - breaking
                          - warning
                          - info
                      changes:
                        type: array
                        items:
                          type: object
                          required:
                            - kind
                            - path
                            - severity
                            - message
                          properties:
                            kind:
                              type: string
                            path:
                              type: string
                            severity:
                              enum:
                                - breaking
                                - warning
                                - info
                            before:
                              type: string
                            after:
                              type: string
                            message:
                              type: string
                      status:
                        enum:
                          - open
                          - accepted
                          - dismissed
                      decidedBy:
                        anyOf:
                          - type: string
                            format: uuid
                          - type: "null"
                      decidedAt:
                        anyOf:
                          - type: string
                            format: date-time
                          - type: "null"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/incidents:
    get:
      summary: List incidents, newest first
      tags:
        - incidents
      parameters:
        - schema:
            enum:
              - open
              - resolved
          in: query
          name: status
          required: false
        - schema:
            type: string
            format: uuid
          in: query
          name: monitorId
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          in: query
          name: limit
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - incidents
                properties:
                  incidents:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - monitorId
                        - monitorName
                        - openedAt
                        - resolvedAt
                        - cause
                        - lastError
                        - failureCount
                      properties:
                        id:
                          type: string
                          format: uuid
                        monitorId:
                          type: string
                          format: uuid
                        monitorName:
                          type: string
                        openedAt:
                          type: string
                          format: date-time
                        resolvedAt:
                          anyOf:
                            - type: string
                              format: date-time
                            - type: "null"
                        cause:
                          type: string
                        lastError:
                          anyOf:
                            - type: string
                            - type: "null"
                        failureCount:
                          type: integer
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
servers:
  - url: /
security:
  - session: []
  - token: []